Implement JWT-based authentication and refactor repository
Added JWT token generation and validation with `JwtService` and integrated it into the application. Refactored repository methods to support condition-based queries. Updated `BotHostedService` to provide authentication tokens, introduced stricter data annotations, and enhanced user and profile models. Additionally, roles-based authorization was implemented in the `UsersController`.

+++ sxkiev/Services/User/IUsersService.cs
@@ -5,7 +5,7 @@ namespace sxkiev.Services.User;
 public interface IUsersService
 {
     Task<IEnumerable<SxKievUser>> GetAllUsersAsync();
-    Task<SxKievUser?> GetUserByIdAsync(int id);
+    Task<SxKievUser?> GetUserByIdAsync(long id);
     Task AddUserAsync(SxKievUser user);
     Task UpdateUserAsync(SxKievUser user);
 }
